overlay-amd64-generic: Add X86 to LLVM_TARGETS
The X86 backend for llvm is needed for llvmpipe on amd64.  The current
version of llvm from arc-toolchain-n automatically includes X86. When
crrev.com/c/698868 goes in, it will switch to a newer llvm ebuild that
needs to be explicitly told to include this backend.

BUG=b:65414758
TEST=betty-pre-cq passes together with crrev.com/c/698868

Change-Id: I4191c70ce230ca8b1cfbf76ff02e30fbd1576615
Reviewed-on: https://chromium-review.googlesource.com/738496
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>