lakitu: fork sys-apps/shadow
So that we could apply private patches. One example is the patch to
allow POSIX useranmes. It'll take a long time to get it accepted by our
upstreams (chromiumos, gentoo, shadow upstream), but we need to get it
released soon.

Package contents are identical to what're in portage-stable now. Also,
mask the copy in portage-stable so we won't accidentally use it when
it's upgraded to a higher version than us.

BUG=b:65754927
TEST=`equery-lakitu w shadow` shows the ebuild from this overlay
RELEASE_NOTE=None

Change-Id: I23a218760fb02d1f3732f37ebbed07fced2c21de
Reviewed-on: https://chromium-review.googlesource.com/740862
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>