baseboard-oak: add USE flag to enable 4.4 kernel with schedtune
This CL adds a USE flag which enables switching the mt8173 chipset
to kernel 4.4. By default, no changes are made and the 3.18 kernel
is used.

Startup scripts are changed to ensure cpusets and schedtune values
are set appropriately (if the 4.4 kernel is selected). These
values are taken from gru/kevin's configuration.

BUG=none
TEST=build for elm with 4.4 kernel:
TEST=export USE="-kernel-3_18 kernel-4_4"
TEST=build_packages --board=elm
TEST=build_image --board=elm
TEST=confirm kernel version is 4.4, CPU governor is sched, and
  chrome/urgent tasks can run on any core, not just big ones:
TEST=uname -a
TEST=cat /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor
TEST=cat /sys/fs/cgroup/cpuset/chrome/urgent/cpus

Change-Id: If59266b53fa20255acefe6ac1208705e47b09288
Signed-off-by: Dave Rodgman <dave.rodgman@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/753736
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>