baseboard-jecht: Also disable power save mode of the audio codec
CL:752324 disables power save mode of the audio controller to work around
an issue with noise due to unstable output on the ALC283 when idle. It
turns out that we also need to disable power save mode of the audio codec
to address the noise issue.

BUG=chromium:775486
TEST=build and boot for guado
  cat /sys/module/snd_hda_intel/parameters/power_save
    => 0
  cat /sys/module/snd_hda_intel/parameters/power_save_controller
    => N
CQ-DEPEND=CL:769160

Change-Id: Ifa6b4fc94d125ec5f036c0a14c04179bfd0d1934
Reviewed-on: https://chromium-review.googlesource.com/766476
Commit-Ready: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>