baseboard-oak: Enable USE=vulkan
We build the vk_glow test when USE=vulkan is enabled.

We attempt to run vk_glow when libvulkan is present.

libvulkan is present on Elm/Hana, but we don't build the
vk_glow test. We then attempt to run the vk_glow test.
Let's fix this.

BUG=chromium:654131
TEST=emerge-elm drm-tests.  vk_glow doesn't work on Elm due
     the fact it can't support XB24 for scanout.  However,
     vk_glow doesn't render anything on Eve either so let's
     build it anyways until we figure what to do regard this
     test.
CQ-DEPEND=CL:818530

Change-Id: I2400ea0041da7c8e80a78d3337058f3603de42cc
Reviewed-on: https://chromium-review.googlesource.com/818383
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>