moblab: Removed dependency to delta_generator
moblab do not use delta_generator anymore and we need to remove it as
moblab's dependency. If we want to keep this dependency through
'delta_generator' IUSE flag, then we have to do the same for all of
delta_generator dependencies which are many and they are large
libraries. This causes a lot of confusion and files leaking into client
devices which should not have been. This is the only dependency to
delta_generator. Once it is removed, we can add delta_generator binary
and its dependecies only for cros_host in the sdk.

moblab still uses devserver for its duties, but not for any delta
generation at this point. So this should be safe to be removed.

BUG=chromium:783883
TEST=tryjob

Change-Id: I0e8b9363eb036c08cb6439d42e81a3dd165db9ed
Reviewed-on: https://chromium-review.googlesource.com/825956
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>