baseboard-cheza: Hack out the TPM for now
Right now we're using baseboard-cheza to boot up MTPs.  Dose ain't got
no TPMs on dem, so for now hack the TPM USE flags out.

When we move to real hardare for cheza we'll have to revert this
patch.  At that time we'll have to decide if we're still going to
support running on MTP.  If so we'll create a new overlay (based on
chipset-qc845) for it.

BUG=chromium:795946
TEST=Boot up and no more respawning of tpm-related scripts

Change-Id: I4eb7356709eea651cc5f37ef530b67f9ff9e77df
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/905804
Reviewed-by: Evan Green <evgreen@google.com>