chromeos-bsp-scarlet: remove old symlink
Somewhere along the way, cherry-picks to R65 got mixed up, and we have
symlinks for both -r7 and -r9 at the moment. The former is useless, and
it makes diffing ToT with R65 a bit noisier.

(I think this was a missed piece in CL:902305, which was trying to
cherry-pick CL:897221, because it was picked *before* CL:902304 which
it was trying to fix up...)

BUG=none
TEST=none

Change-Id: Ia298c4235bb09305486747ca73cd123b77bf2312
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/907491
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>