lakitu: stackdriver: Reuse existing docker network.
Stackdriver docker network will be started when user enable stackdriver
logging or monitoring. This network is a stateful property across
reboot. This CL reuses existing stackdriver docker network if it already
exist.

It is possible that user may have a docker network with the same name.
In which case, we would be using user's network, and that might break
the logging and monitoring agents due to network configuration error.
But this is already best effort.

BUG=b:74013066
TEST=autotest cloud_StackdriverReboot
RELEASE_NOTE=None

Change-Id: I7d099cd2308adbf0c510397b1c7be83484692e97
Reviewed-on: https://chromium-review.googlesource.com/938627
Commit-Ready: Xuewei Zhang <xueweiz@google.com>
Tested-by: Xuewei Zhang <xueweiz@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>