overlay-scarlet: Change usb_min_ac_watts to 11.5
This changes the cutoff for a low power charger to 11.5W. This is to
make sure that 15W chargers connected through a charge-through hub are
not incorrectly identified as low power chargers. We primarily care
about identifying the 7.5W chargers which will be the most common low
power chargers.

BUG=b:71027664
TEST=build scarlet

Change-Id: I20d9447b3c23a09e0f13b983635fb44c8879d570
Reviewed-on: https://chromium-review.googlesource.com/956927
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-by: David Schneider <dnschneid@chromium.org>