Stout: fix the GPIO number for rf-led pin
The rf-led-handler crashed occasionally at boot up time as

  WARNING kernel: [   10.331504] init: rf-led-handler pre-start process
  (1729) terminated with status 2

This patch fixed the crash by correcting the GPIO number.

BUG=chromium:800939
TEST=Build packages and image for stout, cros flash, and reboot.
Turn on wifi or bluetooth in the system tray.
Checked two things:
(1) There should be no crash when grepping the messages log.
    $ grep "rf-led-handler" /var/log/messages
(2) Observe that the green lid led is turned on/off when you
    enable/disable wifi/bluetooth.

Change-Id: Ib033da09c7beb9aba7b93d96e7c655c6b803dcff
Reviewed-on: https://chromium-review.googlesource.com/952944
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>