[moblab] Add explicit dependency on ethtool.
Moblab requires ethtool to detect when the 3.14 kernel and a
realtek driver are running.

If this combination of kernel/driver are running a optimization
has to be switched off or TX packets are dropped.

See comments in files/init/moblab-network-bridge-init.conf for
more details

BUG=chromium:819017
TEST=emerge-guado_moblab chromeos-base/chromeos-bsp-moblab

Change-Id: I2cdc2dc5e735e916899ad7de043e047e1c6dea39
Reviewed-on: https://chromium-review.googlesource.com/956239
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Matt Mallett <mattmallett@chromium.org>