[moblab] Add sleep before trying to detect USB drive.
On the VM it seems like readlink command will hang if called too
early in the boot, this stops the USB drive being mounted and
results in the dev server crashing.

For now add a sleep - however the real issues is that on VM when
cros-disks is started the filesystem still seems unstable.

BUG=chromium:819695
TEST=local tests on moblab vm

Change-Id: I1f6125a62f5e29f6b4f16a46f23f50e6c261c60c
Reviewed-on: https://chromium-review.googlesource.com/957835
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>