lakitu: cherry-pick a patch to containerd
This patch fixes a bug in containerd that it will close a nil chan
when a container init shim process is killed while having an exec attach
to that container.

BUG=b:75975353
TEST=emerge-lakitu containerd
RELEASE_NOTE=Backported a fix for a bug in containerd that killing a
container's shim process while having an exec attch to it will result
in panic error (https://github.com/containerd/containerd/pull/872).

Change-Id: Ieeaf95b75eae39cea7d6ee96d7aae5aea4a7d3d1
Reviewed-on: https://chromium-review.googlesource.com/974800
Commit-Ready: Ke Wu <mikewu@google.com>
Tested-by: Ke Wu <mikewu@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>