lakitu: cloud-init: fix http header value type
cloud-init uses the "requests" python module for HTTP communication.
Older versions of "requests" used to allow (though discouraged) HTTP
header values to be of any type, but since 2.11 they have started to
enforce that headers would have to be strings. As far as I was able to
find, DataSourceGCE.py is the only place where a non-string header is
used, so this CL adds a patch that corrects that.

https://github.com/requests/requests/issues/3477

BUG=chromium:825013
TEST=With this patch, lakitu-release-tryjob passes with the current
version of "requests" as well as a newer one, which CL:984900 is adding.
RELEASE_NOTE=None

Change-Id: I16360edd69ea5ec57a5ae5f1432fe639efef2e18
Reviewed-on: https://chromium-review.googlesource.com/996377
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Keith Haddow <haddowk@chromium.org>