[project-moblab] Install scripts for serving to DUTs.
The devserver code contains two scripts that must occasionally be
installed on DUTs for execution:  `stateful_update` and
`quick-provision`.  To simplify and improve the Autotest code that
does the installation, change the devserver startup to guarantee
that the files are available to be served by the devserver.

CQ-DEPEND=CL:1092235
BUG=None
TEST=ls on the static dir, see the files there.

Change-Id: Ibcc79859be36fb95be133d47a47e47c2886dc5b7
Reviewed-on: https://chromium-review.googlesource.com/1092237
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>