baseboard-reef: Increase rootFS size to 1800 MiB.
reef-paladin was running out of space in the rootFS due to a recent
change to enable CFI when building Chrome which increased the Chrome
binary a bit.  This commit simply bumps up the rootFS size so that
build_RootFilesystemSize may pass again.

BUG=chromium:851897
BRANCH=None
TEST=None

Change-Id: I10b76e8906153902d525f6dd4737a5fd5c0dd9aa
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1097803
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>