chipset-icl: Use kernel-experimental flag
Icelake is a new CPU not supported by 4.14, the support is also not
yet fully upstreamed. The plan is to use Intel's provided code base
for the ICL bringup, which is all the patches in chromeos-4.14 rebased
on top of drm-tip (https://cgit.freedesktop.org/drm-tip), plus some
more patches pulled in from mailing lists. (drm-tip represents the
latest and greatest Intel graphics/drm code, i.e. changes that have been
accepted upstream and either are applied to next or will be applied
soon). Some time in future, we will switch to a new LTS once it becomes
available, which we'll decide later.

We switch to using the kernel-exprimental flag to start picking up code
in kernel/experimental, which will soon be populated with Intel provided
code. At that time, it is desirable to create an intel-grahics/icl
branch and have cros/next/chromeos-experimental point to it.

BUG=b:80267300
TEST=build

Change-Id: Ief17e536e9d6f7a3d44d98232c2b22849bb73399
Reviewed-on: https://chromium-review.googlesource.com/1102990
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>