fizz: Add workaround for 0 fan rpm issue
As discussed in b:109951096, currently fizz board has the issue
that the fan might not start after boot up.

We'd like to add a quick workaround to M67 which checks fan
rpm and recovers the fan if the rpm value is 0 after boot up.

BUG=b:109951096 b:110477076
TEST=Test on a fizz box, the fan is recovered after reboot with the
script.

Change-Id: I357e2298aca815d82a5b03571a2874e5b75c5521
Reviewed-on: https://chromium-review.googlesource.com/1135472
Reviewed-by: Yilun Lin <yllin@chromium.org>
Commit-Queue: Zhongze Hu <frankhu@google.com>
Tested-by: Zhongze Hu <frankhu@google.com>