nyan: Remove cb_legacy_uboot USE flag
This USE flag was originally meant to be used to build a chainloadable
U-Boot payload into the RW_LEGACY section. But it never quite worked in
the end because we didn't get a working U-Boot build ready in time for
Nyan devices. I think(?) it just installed a broken image there.

Now this is leading to weird build errors for some reason, and it
doesn't seem worth the effort to track them down. This never worked
anyway. Just take it out.

BUG=None
TEST=None

Change-Id: I9e3a48392f0701d7a10dee2c081201467e4b7a09
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1125128
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>