Allow experimentation for Instant Tethering on Eve.
Remove default Instant Tethering support on Eve. This is not
intended to actually turn off Instant Tethering for Eve, but
rather will allow the Bluetooth team to run experiments with
Instant Tethering on or off.

This change is targeting M69, and will also need to be cherrypicked
into M68. It will likely be reverted in M70.

BUG=chromium:858884
TEST=None
Previous-Reviewed-on: https://chromium-review.googlesource.com/1128199

Change-Id: I951a82469a7a6fb96b9269dcc4cd0a52c8294532
Reviewed-on: https://chromium-review.googlesource.com/1133769
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Tested-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Ryan Hansberry <hansberry@chromium.org>