Allow experimentation for Instant Tethering on Eve.
Remove default Instant Tethering support on Eve. This is not
intended to actually turn off Instant Tethering for Eve, but
rather will allow the Bluetooth team to run experiments with
Instant Tethering on or off.

This change is targeting M69, and will also need to be cherrypicked
into M68. It will likely be reverted in M70.

BUG=chromium:858884
TEST=None

Change-Id: Ie77b6facff03b0de095b77b07554f8a2f0057b80
Reviewed-on: https://chromium-review.googlesource.com/1128199
Commit-Ready: Ryan Hansberry <hansberry@chromium.org>
Tested-by: Ryan Hansberry <hansberry@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>