Dragonegg: Move ramoops.ecc arg to chipset overlay
Ref to the description of this bug for details:
https://bugs.chromium.org/p/chromium/issues/detail?id=868003

I found that all kernel parameter changes need to be specified in
1 overlay, i.e. cannot be split across multiple overlays. I chose
the chipset overlay to house all parameters for dragon egg because
there is no board specific parameter (so far). In future, if there is
a board specific parameter, we'll have to move all the parameters to
board specific overlay.

BUG=chromium:868003
TEST=check that the kernel arguments are present

Change-Id: Ib11abbc03c0ef23e7782d18f8b3644750d77e081
Reviewed-on: https://chromium-review.googlesource.com/1151900
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>