baseboard-oak: Disable Vulkan on release branch
The Android Nougat CTS has Vulkan failures on Oak. I believe the
failures are largely due to bugs in the CTS itself. Regardless of the
failures' cause, we are not yet prepared to initiate the conformance
process on Oak with Vulkan.

Therefore, globally set USE flag '-vulkan'.

BUG=b:111699930
TEST=`equery-oak use img-ddk` shows vulkan disabled.
TEST=`./build_packages --board=oak`, then confirm no Vulkan files exist
  in /etc/vulkan, /usr/share/vulkan, /usr/lib/,
  /opt/google/containers/android/vendor/.

Change-Id: I62f68ef480139d9833082d894e2e457b4b562fc9
Reviewed-on: https://chromium-review.googlesource.com/1152520
Tested-by: Chad Versace <chadversary@chromium.org>
Trybot-Ready: Chad Versace <chadversary@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>