lakitu: Configure stateful_partition in kernel cmdline
Adds the kernel command line flag 'cos.stateful_dev', used as
'cos.stateful_dev=<device path>:<0|1>'. The 'device path' component
specifies the block device that should be used as the stateful
partition. The component after the colon can be either a '0' or a '1'.
'1' indicates that the stateful partition should be stateless (that is,
mkfs will be run on the stateful partition on every boot) whereas '0'
indicates normal operation. This enables running a system with a read-only
boot disk.

BUG=b:112110620
TEST=Manually changed the kernel cmdline and rebooted
RELEASE_NOTE=None

Change-Id: Ib615ae2c33ba9218fceee0c395b26819be3a0e2f
Reviewed-on: https://chromium-review.googlesource.com/1162903
Commit-Ready: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>