Removed project-freon master and parent references.
It is no longer necessary to maintain project-freon as a separate
entity from the main overlay, so it is being merged into the defaults.
This commit removes instances of freon being referenced as a master
or parent of an overlay.

BUG=chromium:730026
TEST=precq passes
CQ-DEPEND=CL:1165329, CL:1165943, CL:1173411

Change-Id: Ib0f815c3555f299083697c5adb600670069a1e8f
Reviewed-on: https://chromium-review.googlesource.com/1165961
Commit-Ready: Alex Klein <saklein@chromium.org>
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>