nocturne: do not try to adjust keymap on Whiskers touchpad
Whiskers is an USB device containing both keyboard and touchpad, so
naturally input devices get assigned the same vendor and product
numbers, which causes udev to try and adjust keymap not only on
keyboard, but also on touchpad.

Let's adjust the modalias we use in udev rule to also include input
device capability (MSC_SCAN which should only be set for keyboard) so
that it leaves the touchpad alone.

BUG=b:112227695
TEST=Check error messages on nocturne for EVIOCSKEYCODE errors

Change-Id: Ifa6134a672e5a394420dcf53bf3655fa6876cd9d
Reviewed-on: https://chromium-review.googlesource.com/1168584
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>