{amd64,arm,arm64,x32,x86}-generic: move to kernel v4.14
We're supporting a later kernel. Let's move the generic overlays there.

I skip the ppc, and mips overlays, because those are untested and
unmaintained and have been for a while.

I've retained the GCC build for arm32, as I'm not sure 4.14 is up to
snuff for clang builds either.

CQ-DEPEND=CL:1208131
BUG=none
TEST=precq

Change-Id: Ida8839b158ce37da497a30769f6cf7985b135b4d
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1203638
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>