Nami: Set power-supply-full-factor to 97
power-supply-full-factor is the battery charge percentage used by
powerd to decide when to show 'full' to the user. Currently, it's
set to 94%, which is causing the discrepancy between the LED,
controlled by the EC and the taskbar icon, controlled by powerd.

This patch sets it to 97% by removing the configuration from
model.yaml.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,b:80270446
TEST=Deploy chromeos-config and verify powerd sets full_factor=97.

Change-Id: Ic453fa05fc1000e3291aa991716093824be17110
Reviewed-on: https://chromium-review.googlesource.com/1213774
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>