baseboard-octopus: Update SOF topology blob
This ensure that the SOF topology corresponding to the V 0.4 release
gets installed on octopus boards.

repo:      https://github.com/thesofproject/soft
branch:    master
commit ID: d77beaf8bd5260058c00a3fe8f56940eb0962498

NOTE: The previous topology (1.1-octopus) pointed to the same commit,
but the tarball in the chromeos-localmirror was modified, so the most
recent topology might not be getting downloaded. So, we uprev the
topology file and ebuild just to be safe.

BUG=b:112528390, b:110025789
TEST=cros deploy <BOARD> sof-topology and verify that the topology file
     is present at the expected location on-device.

Change-Id: Icbd83631a767094b9828a9c9e216ab1ef719cd73
Reviewed-on: https://chromium-review.googlesource.com/1226278
Commit-Ready: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>