baseboard-*: delete redundant compiler settings
These baseboards inherit from a chipset which sets the same compiler
settings, so once we drop the redundant baseboard settings, we can
drop the now empty make.conf file.

The kunimitsu setting is actually a bit weird.  Setting a more specific
-march flag than the -mtune flag doesn't make sense: the -march flag
will generate code that doesn't work on the -mtune systems.  So drop
the -mtune=generic flag which makes this redundant with the chipset.

We also include eve here as it lacks a baseboard, but has the same
redundancy with its chipset dir.

BUG=chromium:701614
TEST=`portageq-$BOARD envvar CFLAGS CXXFLAGS LDFLAGS` is the same for all boards

Change-Id: Iba99a5d57ac1e873f4948e41e639ab311132ab39
Reviewed-on: https://chromium-review.googlesource.com/1227324
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>