cheza: Fix VPD MAC entries
On rev3 devices the factory programmed the MAC addresses for Bluetooth
and WiFi in the VPD fields 'bluetooth_mac' and 'wifi_mac', however the
correct fields are 'bluetooth_mac0' and 'wifi_mac0'.

Add an init script that updates the VPD to use the correct names if the
incorrect ones are found. The script can be removed again after all rev3
devices went through the update.

BUG=b:117218604
TEST=create 'bluetooth_mac' and 'wifi_mac' entries in the VPD if they
    don't already exist
  reboot the device
  vpd -l
    => entries 'bluetooth_mac0' and 'wifi_mac0' exist and have the same
       values as 'bluetooth_mac' and 'wifi_mac' had earlier. The
       incorrect entries don't exist anymore

Change-Id: Ic89061577a73704ac05bfdfbf3f4fdbcc4731326
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1259522
Reviewed-by: Brian Norris <briannorris@chromium.org>