lakitu: Add option to use sorted list algorithm for choosing stateful device
For CloudSQL distro, we need the ability to choose a specific stateful
device from a collection of devices in a deterministic way.

Suppose one is using an external disk for their stateful device, and they
want to make this disk smaller. One could shut the system down, detach
the stateful device, attach a smaller one, and boot the system.
Alternatively, one could attach a smaller disk, reboot the system, and
the system picks the right stateful device on its own. We are going for
the second approach, and our strategy is to add a configuration option
for choosing the stateful device based on the PD device names of the
devices.

The strategy implemented here entails sorting the relevant device
names (selected by a given prefix) lexicographically and taking the
max device name. This functionality is configured on the kernel command line
and is implemented through an additional systemd service that creates
a symlink to the appropriate device.

TEST=trybot, provisioned read-only VM and successfully resized the stateful
device with only a reboot
BUG=b:117530370
RELEASE_NOTE=None

Change-Id: Id49227fc4657f6df25f9980b6e4c029abf5f77e1
Reviewed-on: https://chromium-review.googlesource.com/1277710
Commit-Ready: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
(cherry picked from commit c6dd3328609886ef5a6c39c2c12fe85be999632a)
Reviewed-on: https://chromium-review.googlesource.com/c/1374071
Reviewed-by: Daniel Wang <wonderfly@google.com>
Commit-Queue: Robert Kolchmeyer <rkolchmeyer@google.com>