overlay-nocturne: set tight_timestamps via udev rule
In the long run, we want tight_timestamps to be setup automatically based on
EC feature flags. While the code to do this is already merged, for already
branched releases of Chrome OS, it is safer to only enable this on specific
boards that need the flag and on which it is actively tested.

As such, this is a temporary change meant to be reverted once we are
comfortable with having CL:1255904 operational on all boards.

Use a udev rule in order to enable this flag for Nocturne.

BUG=b:112111610
TEST=on Nocturne without CL:1255904, emerge and deploy chromeos-bsp-nocturne
     and verify that tight_timestamps is set to 1

Change-Id: Ifb654b8ad96c60fc49f3d1c414b981708115c163
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1287236
Reviewed-by: Alexandru M Stan <amstan@chromium.org>