cheza: Add detachable_ui USE flag
We need this for Cheza. Looks like all other boards define this in their
private overlay, not sure why. depthcharge and vboot build without
private sources, so it probably belongs in here. (It's not like the lack
of a keyboard is a huge secret if you read the public EC sources...)

BUG=None
TEST=None

Change-Id: I57ae4ad173df322bc5c206435f4247bee407ec65
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1295013
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Philip Chen <philipchen@chromium.org>