lakitu: create package makedumpfile.
This package is used to filter & compress & store /proc/vmcore for
kernel dump.

This package does not exist in Gentoo or ChromiumOS upstream. I plan on
keep it in lakitu overlay for now. And we can contribute it to upstream
after we have some usage on it.

BUG=b:115771082
TEST=emerge-lakitu makedumpfile. Will have validation test for kdump.
RELEASE_NOTE=None

Change-Id: Idf6b45f751eea84f8cb371a772b6c6c8a0dd9965
Reviewed-on: https://chromium-review.googlesource.com/1311638
Commit-Ready: Xuewei Zhang <xueweiz@google.com>
Tested-by: Xuewei Zhang <xueweiz@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>