lakitu: add systemd units and kdump_helper to implement kdump on COS.
This CL mainly contains three things:
1. kdump_helper script
This script is used to do all kdump utilities, such as status checking,
enable/disable feature, load the kdump kernel, or take the kernel dump.

2. kdump-load-kernel.service
This unit is started when "crashkernel" kernel parameter is set, and
will load the kdump kernel into reserved zone if kdump is enabled.

3. kdump-save-dump.service
This is the systemd unit to boot into upon crash. The default
multi-user.target is not necessary, since the only thing we will be
doing is to save the kernel dump.

BUG=b:115771082
TEST=TODO: validation test + VmTest
RELEASE_NOTE=None

Change-Id: If76323b9b8ef6f407d96056cfa0112584e87f123
Reviewed-on: https://chromium-review.googlesource.com/1311640
Commit-Ready: Xuewei Zhang <xueweiz@google.com>
Tested-by: Xuewei Zhang <xueweiz@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>
Reviewed-by: Daniel Wang <wonderfly@google.com>