amd64-generic: Expand kernel partition to 64 MB
We're running out of disk space for the kernel image when we enable all
the zillions of debug options that we have. For example, an
amd64-generic kernel with the 'debug' and 'ubsan' use flags enabled is
26 MB. It's suggested that we increase the size to 64 MB to be safe for
the foreseeable future. Let's do that, and also increase the EFI
partition to be twice the size of the kernel partition because we copy
the vmlinuz file there on this architecture.

BUG=chromium:894237
TEST=Build image and see partitions 2, 4 and 12 are the size intended

Change-Id: I63b4afd23243cc1777fdb318e5f57ca86e71d9c6
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1318213
Reviewed-by: Douglas Anderson <dianders@chromium.org>