lakitu: Relax the ipv6 firewall rules a bit
Our original firewall for ipv6 was to drop all. That's not going to work
if we want to do ipv6. This CL updates it to match what we have for
ipv4: disallows virtually everything inbound except for SSH, and DHCPv6.

This will be a noop on VMs that don't have ipv6 turned on, or don't have
a router that does ipv6 so we can have it in the lakitu overlay.

BUG=b:74356284
TEST=When ipv6 is turned on, with these firewall rules, the VM is able
    to ipv6 and DHCPv6.
RELEASE_NOTE=None

Change-Id: I59c9f2607ce86e768ed3e4bd34cea5a94dfb27b9
Reviewed-on: https://chromium-review.googlesource.com/1321355
Commit-Ready: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Aditya Kali <adityakali@google.com>