lakitu: Update lakitu-custom-locales with glibc 2.27
The new glibc version requires that "category" be one of "posix:1993",
"i18n:2004", and "i18n:2012". Change our locale def to match. Also,
append `|| die` to `src_install` so that the build fails when the
installation of the locale fails.

https://sourceware.org/git/?p=glibc.git;a=commitdiff;h=900f59f084bfe35cb389bbe0dc464413a1a38e90.

While looking at this I realized that we were using the `localedef` in
the SDK to install the locale on the board. While there is the
theoretical hazard that the compiled locale archive may not be
compatible with the glibc installed on the board, according to vapier@
"glibc has been moving to make the locale archive format arch
independent (by forcing fields to be aligned the same regardless of
arch/bitsize).  so with newer versions, this isn't as dicey as it used
to be".

Based on that, I am converting the patch into a source file that we can
use directly, eliminating the need to compile the full glibc.

BUG=b:119335610
TEST=platform_locale passes again
RELEASE_NOTE=None

Change-Id: I2e70c658425d515b020ddc9f45a9dfe4894b20b3
Reviewed-on: https://chromium-review.googlesource.com/c/1331152
Reviewed-by: Daniel Wang <wonderfly@google.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Commit-Queue: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>