octopus: Add $num-mic variable to cros_config.
This change introduces a new num-mic field that's used instead of mic
config for various SKUs. This is done with the intent of introducing
"num-mic" and "num-camera" fields to the schema, which will both be set
to {{$num-mic}} today. This lets various consumers (audio configuration,
ARC++ media profiles, ARC++ hardware features) correctly configure the
system for multiple cameras/mics, while keeping the schema free in the
future to have different numbers of microphones and cameras.

For handling camera, A more ambitious change would be to have the schema
add entries like:
  front_camera: <vid:pid>
  rear_camera: <vid:pid>
...but I believe that for today, simply knowing the number will be
enough.

BUG=None
TEST=emerge-octopus chromeos-config-bsp-octopus chromeos-config

Change-Id: Ia6913badc0a3c0a4efeba1bcb3a904622fef22d6
Reviewed-on: https://chromium-review.googlesource.com/1335996
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>