baseboard-nami: adjust suitable media perf range from Vayne and Akali
Nami family is uni-built image we are not able to have specific arc-codec package for specific boards.
That is, there is unified media_codecs_performance.xml file among Nami family and it needs to cover the
performance metrics from the lowest SKU to the highest.

This change adjusts all perf ranges with target to be compatible for the lowest SKU (Akali) and the
highest (Vayne).

BUG=b:119601141
BUG=b:119651252
TEST=build arc-codec package for Nami

Change-Id: Ia45023f1b85fac08def845fe52541c315217a3f3
Reviewed-on: https://chromium-review.googlesource.com/1341690
Commit-Ready: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>