chromeos-config-bsp-octopus: Add camera-count
Now that this is added to the schema, we add a camera-count entry to the
model so that the field can be used.

Currently, there must be a microphone available near each camera on the
system for recording audio for video streams. Given this, we set the
values to be equivalent on Octopus (assuming the number of cameras is
equivalent to the number of locations that microphones are present at).

If requirements change in the future, the definition of camera count
would change, and be overriden in this file.

BUG=b:118248952
CQ-DEPEND=CL:1355504
TEST="cros_config_host -c $OCT_YAML -m yorp get /camera count" = 2
TEST="cros_config_host -c $OCT_YAML -m fleex get /camera count" = 1

Change-Id: I6f2a274f5a9e7404fb41bd0446cfb1b76c76d093
Reviewed-on: https://chromium-review.googlesource.com/1355426
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>
(cherry picked from commit 2473f7be86d2f8e91a3e4bc2cdd584ea83e150d5)
Reviewed-on: https://chromium-review.googlesource.com/c/1368008
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Commit-Queue: Justin TerAvest <teravest@chromium.org>
Trybot-Ready: Justin TerAvest <teravest@chromium.org>