coral: Fix issue with dptf-dv param
Originally, it didn't look like dptf-dv was used anywhere, but it ended
up embedded in a Chrome binary payload that does end up on the target.
See bug for context.

Adding it back to diff'ing support since this is a valid parameter.

BUG=b:120088670
TEST=local build/test on coral and diff testing
CQ-DEPEND=CL:1352228

Change-Id: I8bb682d140269e4ae57117680adc658b59f65aea
Reviewed-on: https://chromium-review.googlesource.com/1352372
Commit-Ready: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>