coral: Migrate dptf files to cros_config
Removing the dependency on the device-specific inheritance hierarchy and
switched to using native cros_config to model dptf file dependencies and
installation.

There was a bug in the file path, but it was a straight port of the bug
from the device-tree impl (since it was never used before, it was never
found).

BUG=b:119679625
TEST=see diff test results
CQ-DEPEND=CL:1352224

Change-Id: Iddee03de586888ed3460a95855f4999de18b45c2
Reviewed-on: https://chromium-review.googlesource.com/1352377
Commit-Ready: C Shapiro <shapiroc@chromium.org>
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>