octopus: Set min-visible-backlight-level to 960
A panel used on the Fleex project requires a minimum 5% PWM to work
correctly. The maximum brightness for the intel_backlight driver is
19200, giving a value of 960.

Even though powerd is using a step of 6.25%, we use 5% in this change if
the number of backlight steps is changed in the future.

I built this change, confirmed that the change was picked up by powerd,
and that the brightness keys on the keyboard functioned correctly.

BUG=b:120062768
TEST=emerge, cros deploy, change brightness and check logs

overlay-octopus/chromeos-base/chromeos-config-bsp-octopus/files/model.yaml

Change-Id: If40ccf0cc8d7fddbeb680829c10a1496b2dc240a
Reviewed-on: https://chromium-review.googlesource.com/1359111
Commit-Ready: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>